+++ packages/bpk-component-card/src/BpkCard.js
@@ -4,9 +4,10 @@ import './bpk-card.scss';
 
 const BpkCard = (props) => {
   const classNames = ['bpk-card'];
-  const { children, href, padded, ...rest } = props;
+  const { children, className, href, padded, ...rest } = props;
 
   if (padded) { classNames.push('bpk-card--padded'); }
+  if (className) { classNames.push(className); }
 
   const classNameFinal = classNames.join(' ');
 
@@ -19,11 +20,13 @@ const BpkCard = (props) => {
 
 BpkCard.propTypes = {
   children: PropTypes.node.isRequired,
+  className: PropTypes.string,
   href: PropTypes.string,
   padded: PropTypes.bool,
 };
 
 BpkCard.defaultProps = {
+  className: null,
   href: null,
   padded: true,
 };