+++ packages/bpk-component-ticket/index.js
@@ -1,0 +1,3 @@
+import BpkTicket from './src/BpkTicket';
+
+export default BpkTicket;