+++ packages/bpk-component-ticket/src/BpkTicket-test.js
@@ -1,44 +1,54 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
-import BpkCard from './BpkCard';
+import BpkTicket from './BpkTicket';
 
-describe('BpkCard', () => {
+describe('BpkTicket', () => {
   it('should render correctly', () => {
     const tree = renderer.create(
-      <BpkCard>
+      <BpkTicket stub="Ticket stub">
         Lorem ipsum dolor sit amet, consectetuer adipiscing elit. Aenean commodo ligula eget dolor. Aenean massa. Cum
         sociis natoque penatibus et magnis dis parturient montes, nascetur ridiculus mus.
-      </BpkCard>,
+      </BpkTicket>,
     ).toJSON();
     expect(tree).toMatchSnapshot();
   });
 
   it('should render correctly with an "href" attribute', () => {
     const tree = renderer.create(
-      <BpkCard href="//www.skyscanner.net">
+      <BpkTicket stub="Ticket stub" href="//www.skyscanner.net">
         Lorem ipsum dolor sit amet, consectetuer adipiscing elit. Aenean commodo ligula eget dolor. Aenean massa. Cum
         sociis natoque penatibus et magnis dis parturient montes, nascetur ridiculus mus.
-      </BpkCard>,
+      </BpkTicket>,
     ).toJSON();
     expect(tree).toMatchSnapshot();
   });
 
   it('should render correctly with "padded" attribute equal to "false"', () => {
     const tree = renderer.create(
-      <BpkCard padded={false}>
+      <BpkTicket stub="Ticket stub" padded={false}>
         Lorem ipsum dolor sit amet, consectetuer adipiscing elit. Aenean commodo ligula eget dolor. Aenean massa. Cum
         sociis natoque penatibus et magnis dis parturient montes, nascetur ridiculus mus.
-      </BpkCard>,
+      </BpkTicket>,
+    ).toJSON();
+    expect(tree).toMatchSnapshot();
+  });
+
+  it('should render correctly with a "vertical" attribute', () => {
+    const tree = renderer.create(
+      <BpkTicket stub="Ticket stub" vertical>
+        Lorem ipsum dolor sit amet, consectetuer adipiscing elit. Aenean commodo ligula eget dolor. Aenean massa. Cum
+        sociis natoque penatibus et magnis dis parturient montes, nascetur ridiculus mus.
+      </BpkTicket>,
     ).toJSON();
     expect(tree).toMatchSnapshot();
   });
 
   it('should render correctly with a "className" attribute', () => {
     const tree = renderer.create(
-      <BpkCard className="custom-class">
+      <BpkTicket stub="Ticket stub" className="custom-class">
         Lorem ipsum dolor sit amet, consectetuer adipiscing elit. Aenean commodo ligula eget dolor. Aenean massa. Cum
         sociis natoque penatibus et magnis dis parturient montes, nascetur ridiculus mus.
-      </BpkCard>,
+      </BpkTicket>,
     ).toJSON();
     expect(tree).toMatchSnapshot();
   });