+++ packages/bpk-docs/src/layouts/DocsLayout/DocsLayout.js
@@ -42,6 +42,7 @@ const links = [
       { route: routes.ACCORDIONS, children: 'Accordions' },
       { route: routes.NUDGERS, children: 'Nudgers' },
       { route: routes.PROGRESS, children: 'Progress bars' },
+      { route: routes.TICKETS, children: 'Tickets' },
       { route: null, children: 'Pagination' },
       { route: null, children: 'Slider' },
       { route: null, children: 'Carousel' },